ranobe rf fix payment status

+++ app/src/main/java/ru/profapp/ranobe/network/dto/ranoberfDTO/RfUserGson.kt
@@ -26,7 +26,7 @@ data class ResultUser(
         @SerializedName("paymentContinue")
         var paymentContinue: Long? = null,
         @SerializedName("donates")
-        var donates: Int? = null,
+        var donates: Double? = null,
         @SerializedName("checkerBlock")
         var checkerBlock: Any? = null
